crwal4ai/requirements.txt
# Sahibinden Crawler Dependencies
crawl4ai[all]>=0.7.8
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0

# Supabase Client
supabase>=2.0.0

# Undetected ChromeDriver - Cloudflare bypass
undetected-chromedriver>=3.5.0
selenium>=4.15.0

# 2Captcha Turnstile Solver
2captcha-python>=1.2.0

# FastAPI Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
